Moved the logic to discover subscribers and execute handlers into SubscriberServiceHelper so it is reusable for a StatefulService. Removed the Subscribe and Unsubscribe endpoints.
